refactor(sip): extract the static SipMessage pool into SipMessagePool
Pure move of the #53/#101(A)/#101(E) pool machinery out of
RequestsHandler.cpp: the leaf mutex, bounded heap fallback, deleter,
PoolPressure/logPoolExhausted, and both getMessageFromPool overloads.
RequestsHandler keeps the two public statics as one-line forwarders;
SipMessageFactory and the tests are unchanged. allocateVirtualPeer
stays in the engine (instance-owned pool under _mutex) and uses the
shared logger via sipmsgpool::.

The lock-ordering essay moves with the code; two references were
recontextualized (_messagePool -> the pool, _mutex -> RequestsHandler's
_mutex) so they don't dangle in the new file. No logic, constant,
memory order, or exception path changed.

docs/REALITY_CHECK.md's line refs were already stale before this and
are deliberately untouched. ISSUES.md/CHANGELOG.md are history.

Host suite 309/309 (excluding one pre-existing flaky test, see PR).
Firmware compiles clean for SIP_TRANSPORT=eth. Callgraph guard clean.
