+++ main/CMakeLists.txt
@@ -90,6 +90,7 @@ idf_component_register(SRCS ${MAIN_SRC}
                             "../src/Helpers/SipSecretStore.cpp"
                             "../src/Helpers/ArpLookup.cpp"
                             "../src/Helpers/OtaUpdater.cpp"
+                            "../src/SIP/SipMessagePool.cpp"
                             "../src/SIP/RequestsHandler.cpp"
                             "../src/SIP/TransactionLayer.cpp"
                             "../src/SIP/Registrar.cpp"
